+++ spec/server_spec.rb
@@ -85,7 +85,7 @@ describe 'openstack-dashboard::server' do
 
       it 'has proper owner' do
         expect(file.owner).to eq('root')
-        expect(file.group).to eq('root')
+        expect(file.group).to eq('horizon')
       end
 
       it 'has proper modes' do